*Changed the way the flags are used for ALE. From now on, the only variable in the input is aleType, which might take different values:         0/ no ALE         1/ Rigid body motion -> aleOn = 1, aleRigid = 1         2/ Imposed inner mesh motion -> aleOn = 1, aleImposedMeshMotion = 1         3/ (to be implemented...)
*Changed the name globalMeshVelocity to globalRigidVelocity, which I think is
more appropiate

*Created global array uMesh and dispMesh. This ones will always exist, though
they will take a value different than zero only in the specific cases 1, 2, 3...
above.

*Started passing down variable uMesh until AsIGMR. At AsIGMR I've created the
variable uMeshl. This is send to e3ivar which returns uMesh1, uMesh2, uMesh3,
which are the 3 coordinates of uMeshl evaluated at the integration point.
From this point, uMesh1, uMesh2, uMesh3 can be send to the rest of the subroutines
The gathering has been done using subroutine localy, which seems to
be consistent with what we need.

*removed old unnecessary flags

* all Chris tests passed